first major refactor step: add `elements` directory, which are reserved for presentational components; groups elements/components by function in separate directories
