+++ src/elements/line-chart/index.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import XAxisTick from './x-axis-tick';
 import {
+  ResponsiveContainer,
   AreaChart,
   Area,
   XAxis,
@@ -8,45 +9,43 @@ import {
   Tooltip,
 } from 'recharts';
 
-import '../../css/index.css';
-
 import { formatMoney } from '../../utils/formatMoney';
 
-class Chart extends Component {
+class LineChart extends Component {
   render() {
     const newArr = this.props.graphData;
     return (
-      <div className="chartContainer">
-          <AreaChart
-            data={ newArr }
-            margin={{ top: 5, right: 30, left: 20, bottom: 5 }}
-          >
-            <XAxis
-              margin={{ bottom: 100 }}
-              padding={{ top: 20 }}
-              interval={ 0 }
-              tick={ <XAxisTick {...this.props}/> }
-              dataKey='age'
-            />
-            <YAxis
-              tickFormatter={ money => '$' + formatMoney(+money, 0)}
-              dataKey='savings'
-            />
-            <Area
-              type="monotone"
-              dataKey="savings"
-              stroke='#0074D9'
-              fill='#0074D9'
-            />
-            <Tooltip
-              label={ 'savings' }
-              labelFormatter={(age) => (`Age: ${age}`)}
-              formatter={(money) => (`$${formatMoney(+money, 0)}`)}
-            />
-          </AreaChart>
-      </div>
+      <ResponsiveContainer width={'75%'} height={'90%'}>
+        <AreaChart
+          data={ newArr }
+          margin={{ top: 5, right: 30, left: 20, bottom: 5 }}
+        >
+          <XAxis
+            margin={{ bottom: 100 }}
+            padding={{ top: 20 }}
+            interval={ 0 }
+            tick={ <XAxisTick {...this.props}/> }
+            dataKey='age'
+          />
+          <YAxis
+            tickFormatter={ money => '$' + formatMoney(+money, 0)}
+            dataKey='savings'
+          />
+          <Area
+            type="monotone"
+            dataKey="savings"
+            stroke='#0074D9'
+            fill='#0074D9'
+          />
+          <Tooltip
+            label={ 'savings' }
+            labelFormatter={(age) => (`Age: ${age}`)}
+            formatter={(money) => (`$${formatMoney(+money, 0)}`)}
+          />
+        </AreaChart>
+      </ResponsiveContainer>
     );
   }
 }
 
-export default Chart;
+export default LineChart;
